Zeno record

+++ zenovis/xinxinoptix/DeflMatShader.cu
@@ -159,6 +159,7 @@ static __inline__ __device__ float3 sphereUV(float3 &direction) {
 
 extern "C" __global__ void __anyhit__shadow_cutout()
 {
+
     const OptixTraversableHandle gas = optixGetGASTraversableHandle();
     const uint           sbtGASIndex = optixGetSbtGASIndex();
     const uint              prim_idx = optixGetPrimitiveIndex();
@@ -204,9 +205,9 @@ extern "C" __global__ void __anyhit__shadow_cutout()
 
     unsigned short isLight = 0;
 #else
-    int inst_idx2 = optixGetInstanceIndex();
-    int inst_idx = rt_data->meshIdxs[inst_idx2];
-    int vert_idx_offset = (inst_idx * TRI_PER_MESH + prim_idx)*3;
+    size_t inst_idx2 = optixGetInstanceIndex();
+    size_t inst_idx = rt_data->meshIdxs[inst_idx2];
+    size_t vert_idx_offset = (inst_idx * TRI_PER_MESH + prim_idx)*3;
 
     float m16[16];
     m16[12]=0; m16[13]=0; m16[14]=0; m16[15]=1;
@@ -469,9 +470,11 @@ extern "C" __global__ void __closesthit__radiance()
     }
     prd->test_distance = false;
 
+
     const OptixTraversableHandle gas = optixGetGASTraversableHandle();
     const uint           sbtGASIndex = optixGetSbtGASIndex();
     const uint              prim_idx = optixGetPrimitiveIndex();
+
 
     const float3 ray_orig = optixGetWorldRayOrigin();
     const float3 ray_dir  = optixGetWorldRayDirection();
@@ -513,9 +516,9 @@ extern "C" __global__ void __closesthit__radiance()
 
 #else
 
-    int inst_idx2 = optixGetInstanceIndex();
-    int inst_idx = rt_data->meshIdxs[inst_idx2];
-    int vert_idx_offset = (inst_idx * TRI_PER_MESH + prim_idx)*3;
+    size_t inst_idx2 = optixGetInstanceIndex();
+    size_t inst_idx = rt_data->meshIdxs[inst_idx2];
+    size_t vert_idx_offset = (inst_idx * TRI_PER_MESH + prim_idx)*3;
 
     float m16[16];
     m16[12]=0; m16[13]=0; m16[14]=0; m16[15]=1;
